# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: InvestFlow Investment Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /strategies:
    get:
      operationId: listStrategies
      tags: [strategies]
      summary: List all available strategies
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [active, paused, closed]
        - in: query
          name: riskLevel
          schema:
            type: string
            enum: [low, medium, high]
        - in: query
          name: category
          schema:
            type: string
      responses:
        "200":
          description: Strategies list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Strategy"
  /strategies/{id}:
    get:
      operationId: getStrategy
      tags: [strategies]
      summary: Get strategy details
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Strategy details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StrategyDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /deposits:
    get:
      operationId: listDeposits
      tags: [deposits]
      summary: List my deposits
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, confirmed, failed, cancelled]
        - in: query
          name: method
          schema:
            type: string
      responses:
        "200":
          description: Deposits list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deposit"

    post:
      operationId: createDeposit
      tags: [deposits]
      summary: Create deposit request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositInput"
      responses:
        "201":
          description: Deposit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
...
  /withdrawals:
    get:
      operationId: listWithdrawals
      tags: [withdrawals]
      summary: List my withdrawals
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, processing, completed, rejected]
      responses:
        "200":
          description: Withdrawals list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Withdrawal"

    post:
      operationId: requestWithdrawal
      tags: [withdrawals]
      summary: Request withdrawal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalInput"
      responses:
        "201":
          description: Withdrawal requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
        "400":
          description: Insufficient balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/dashboard:
    get:
      operationId: getAdminDashboard
      tags: [admin]
      summary: Admin dashboard stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDashboard"
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [active, suspended, pending]
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Users page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPage"
  /admin/users/{id}:
    get:
      operationId: adminGetUser
      tags: [admin]
      summary: Get user details
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"

    patch:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Update user
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/strategies:
    get:
      operationId: adminListStrategies
      tags: [admin]
      summary: Admin list strategies
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Strategies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Strategy"

    post:
      operationId: adminCreateStrategy
      tags: [admin]
      summary: Create strategy
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StrategyInput"
      responses:
        "201":
          description: Created strategy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
  /admin/strategies/{id}:
    patch:
      operationId: adminUpdateStrategy
      tags: [admin]
      summary: Update strategy
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StrategyUpdate"
      responses:
        "200":
          description: Updated strategy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"

    delete:
      operationId: adminDeleteStrategy
      tags: [admin]
      summary: Delete strategy
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/strategies/{id}/yield:
    post:
      operationId: adminApplyYield
      tags: [admin]
      summary: Apply yield/rentabilidade to a strategy
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/YieldInput"
      responses:
        "200":
          description: Yield applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/deposits:
    get:
      operationId: adminListDeposits
      tags: [admin]
      summary: Admin list all deposits
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Deposits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositPage"
  /admin/deposits/{id}/confirm:
    post:
      operationId: adminConfirmDeposit
      tags: [admin]
      summary: Confirm deposit
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmDepositInput"
      responses:
        "200":
          description: Confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
  /admin/withdrawals:
    get:
      operationId: adminListWithdrawals
      tags: [admin]
      summary: Admin list all withdrawals
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Withdrawals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalPage"
  /admin/withdrawals/{id}/approve:
    post:
      operationId: adminApproveWithdrawal
      tags: [admin]
      summary: Approve or reject withdrawal
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveWithdrawalInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
  /admin/settings:
    get:
      operationId: getAdminSettings
      tags: [admin]
      summary: Get platform settings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettings"

    patch:
      operationId: updateAdminSettings
      tags: [admin]
      summary: Update platform settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlatformSettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettings"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Strategy:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        riskLevel:
          type: string
          enum: [low, medium, high]
        category:
          type: string
        minInvestment:
          type: number
        totalShares:
          type: integer
        availableShares:
          type: integer
        sharePrice:
          type: number
        aum:
          type: number
        maxDrawdown:
          type: number
        totalReturnPct:
          type: number
        monthlyReturnPct:
          type: number
        status:
          type: string
          enum: [active, paused, closed]
        startDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
      required: [id, name, description, riskLevel, category, minInvestment, totalShares, availableShares, sharePrice, aum, maxDrawdown, totalReturnPct, monthlyReturnPct, status, startDate, createdAt]
    StrategyDetail:
      allOf:
        - $ref: "#/components/schemas/Strategy"
        - type: object
          properties:
            performanceHistory:
              type: array
              items:
                $ref: "#/components/schemas/PerformancePoint"
            userPosition:
              $ref: "#/components/schemas/Position"
          required: [performanceHistory]
...
    DepositInput:
      type: object
      required: [method, amount]
      properties:
        method:
          type: string
          enum: [pix, usdt_bep20, bitcoin, usdc, bnb]
        amount:
          type: number
          minimum: 1
...